refactor(reborn): centralize secret store wiring
